app/requirements.txt
django-cors-headers==3.1.1    
django-crispy-forms==1.8.0    
django-filter==2.2.0    
django-filters==0.2.1    
djangorestframework==3.10.3
